Rename ResourceAttributeMap to Model.
This also changes parse_from_resource/1 to return a Model struct,
which contains a list of attribute structs.
