+++ lib/braise/attribute.ex
@@ -1,0 +1,3 @@
+defmodule Braise.Attribute do
+  defstruct [:name, :type, :format]
+end
